WIP: Create new Orbit object, make API changes
The maneuver application does not work yet and the docs
are now outdated.
