+++ src/poliastro/twobody/equinoctial.py
@@ -34,9 +34,8 @@ def mee2coe(p, f, g, h, k, L):
 
 
 class ModifiedEquinoctialState(BaseState):
-    def __init__(self, attractor, p, f, g, h, k, L,
-                 epoch):
-        super(ModifiedEquinoctialState, self).__init__(attractor, epoch)
+    def __init__(self, attractor, p, f, g, h, k, L):
+        super(ModifiedEquinoctialState, self).__init__(attractor)
         self._p = p
         self._f = f
         self._g = g
