+++ src/poliastro/twobody/tests/test_core.py
@@ -2,14 +2,18 @@
 import pytest
 
 import numpy as np
-from numpy.testing import (assert_equal, assert_almost_equal,
+
+from numpy.testing import (assert_almost_equal,
                            assert_array_almost_equal)
 
 from astropy import units as u
-from astropy import time
 
 from poliastro.bodies import Sun, Earth
 from poliastro.twobody import State
+
+from poliastro.twobody.propagation import cowell
+
+from poliastro.util import norm
 
 
 def test_state_has_attractor_given_in_constructor():
@@ -18,16 +22,6 @@ def test_state_has_attractor_given_in_constructor():
     _a = 1.0 * u.deg  # Unused angle
     ss = State.from_classical(Sun, _d, _, _a, _a, _a, _a)
     assert ss.attractor == Sun
-
-
-def test_default_time_for_new_state():
-    _d = 1.0 * u.AU  # Unused distance
-    _ = 0.5 * u.one  # Unused dimensionless value
-    _a = 1.0 * u.deg  # Unused angle
-    _body = Sun  # Unused body
-    expected_epoch = time.Time("J2000", scale='utc')
-    ss = State.from_classical(_body, _d, _, _a, _a, _a, _a)
-    assert ss.epoch == expected_epoch
 
 
 def test_state_has_elements_given_in_constructor():
@@ -176,31 +170,6 @@ def test_convert_from_coe_to_rv():
     assert_almost_equal(nu.to(u.deg).value, 92.335, decimal=2)
 
 
-def test_apply_zero_maneuver_returns_equal_state():
-    _d = 1.0 * u.AU  # Unused distance
-    _ = 0.5 * u.one  # Unused dimensionless value
-    _a = 1.0 * u.deg  # Unused angle
-    ss = State.from_classical(Sun, _d, _, _a, _a, _a, _a)
-    dt = 0 * u.s
-    dv = [0, 0, 0] * u.km / u.s
-    ss_new = ss.apply_maneuver([(dt, dv)])
-    assert_almost_equal(ss_new.r.to(u.km).value,
-                        ss.r.to(u.km).value)
-    assert_almost_equal(ss_new.v.to(u.km / u.s).value,
-                        ss.v.to(u.km / u.s).value)
-
-
-def test_apply_maneuver_changes_epoch():
-    _d = 1.0 * u.AU  # Unused distance
-    _ = 0.5 * u.one  # Unused dimensionless value
-    _a = 1.0 * u.deg  # Unused angle
-    ss = State.from_classical(Sun, _d, _, _a, _a, _a, _a)
-    dt = 1 * u.h
-    dv = [0, 0, 0] * u.km / u.s
-    ss_new = ss.apply_maneuver([(dt, dv)])
-    assert ss_new.epoch == ss.epoch + dt
-
-
 def test_perifocal_points_to_perigee():
     _d = 1.0 * u.AU  # Unused distance
     _ = 0.5 * u.one  # Unused dimensionless value
@@ -231,3 +200,54 @@ def test_pqw_returns_dimensionless():
     assert p.unit == u.one
     assert q.unit == u.one
     assert w.unit == u.one
+
+
+def test_cowell_propagation_with_zero_acceleration_equals_kepler():
+    # Data from Vallado, example 2.4
+    k = Earth.k.to(u.km**3 / u.s**2).value
+
+    r0 = np.array([1131.340, -2282.343, 6672.423])  # km
+    v0 = np.array([-5.64305, 4.30333, 2.42879])  # km/s
+    tof = 40 * 60.0  # s
+
+    expected_r = np.array([-4219.7527, 4363.0292, -3958.7666])
+    expected_v = np.array([3.689866, -1.916735, -6.112511])
+
+    r, v = cowell(k, r0, v0, tof, None)
+
+    assert_array_almost_equal(r, expected_r, decimal=1)
+    assert_array_almost_equal(v, expected_v, decimal=4)
+
+
+def test_cowell_propagation_circle_to_circle():
+    # From [Edelbaum, 1961]
+    accel = 1e-7
+
+    def constant_accel(t0, u, k):
+        v = u[3:]
+        norm_v = (v[0]**2 + v[1]**2 + v[2]**2)**.5
+        return accel * v / norm_v
+
+    ss = State.circular(Earth, 500 * u.km)
+    tof = 20 * ss.period
+
+    r0, v0 = ss.rv()
+    k = ss.attractor.k
+
+    r, v = cowell(k.to(u.km**3 / u.s**2).value,
+                  r0.to(u.km).value,
+                  v0.to(u.km / u.s).value,
+                  tof.to(u.s).value,
+                  constant_accel)
+
+    ss_final = State.from_vectors(Earth,
+                                  r * u.km,
+                                  v * u.km / u.s)
+
+    da_a0 = (ss_final.a - ss.a) / ss.a
+    dv_v0 = abs(norm(ss_final.v) - norm(ss.v)) / norm(ss.v)
+    assert_almost_equal(da_a0.value, 2 * dv_v0.value, decimal=4)
+
+    dv = abs(norm(ss_final.v) - norm(ss.v))
+    accel_dt = accel * u.km / u.s**2 * tof
+    assert_almost_equal(dv.value, accel_dt.value, decimal=4)
